chore: Upgrade packages to fix vulnerabilities (#531)
* chore: Upgrade many dependencies to resolve minimist vulnerability

* chore: Upgrade cypress

* chore: Add minimist resolutions for optimist and commitizen

* chore: Yarn dedupe

* chore: Update react-scripts in yarn and update resolutions

* feat(tooltip): Refactor to a simpler API (#528)

* Add unit tests
* Add Cypress specifications
* Add visual tests
* Upgrade Popper to v2 and use canvas-kit Popper component instead of materia-ui's

This change makes it easier to create accessible tooltips. It is now possible to do this:

```tsx
<Tooltip title="Close">
  <IconButton variant={IconButton.Variant.Circle} icon={xIcon} aria-label="Close" />
</Tooltip>
```

BREAKING CHANGE: Tooltip API was updated to make it easier to consume.
The original `Tooltip` did little more than add a `role="tooltip"` to a styled component. The original tooltip is now exported as `TooltipContainer` to make it easier to migrate without rewriting all tooltips. Imports will have to be updated to use the old API:
**Before:**
```ts
import { Tooltip } from '@workday/canvas-kit-react-tooltip'
```
**After:**
```ts
import { TooltipContainer as Tooltip } from '@workday/canvas-kit-react-tooltip'
```
Also with this change, the tooltip no longer gets the role `tooltip` and must be added manually.

BREAKING CHANGE: `Popper` was changed to a Functional Component with a forwarded ref. If you passed a `ref` object to `Popper` before, it will now point to the element rather than the `Popper` instance.

* chore: Remove tslint deps added by rebase

* build: Remove react-scripts dependency

* build: Fix postcss/scss issues in webpack config

* chore: Simplify resolutions

* chore: Revert rollup-plugin-terser upgrade

* chore: Change http-server usage to npx so we can remove dep

* chore: Upgrade commitizen to avoid minimist resolution

* build: Automate css copyright date

* fix: Remove unused import

Co-authored-by: Nicholas Boll <nicholas.boll@gmail.com>