feat: Add a Popup Stack manager to Canvas Kit (#670)
Fixes #182

Adds a popup stack manager to Canvas Kit. This popup stack is framework agnostic and meant to be used globally throughout an entire application stack. It takes care to normalize z-index values, escape key handling, and click outside handling.

Components have been modified to use this stack so that escape key and click outside listeners work with other popups in the stack. This is needed for accessibility requirements and so that Canvas Kit can work with other popup systems (so long as those popup systems consume this stack).

BREAKING CHANGES:
* Popper has been moved to the Popup module and out of common
* `closeOnEscape` has been removed from the Modal component as it is no longer needed