docs: Re-order content and fix links [skip release] (#1286)
A recent contribution demonstrated we could outline first steps in a better order. In particular, we should mention the fork and branch strategy for PRs earlier in the process.

I also noticed a couple links that do not work, and the absence of `CODE_OF_CONDUCT.md` and `CONTRIBUTING.md` at the root of our project, which some people look for when browsing GitHub. I moved the `CODE_OF_CONDUCT.md` from its currently location back to root, and added a `CONTRIBUTING.md` file that links to the Storybook page.

[category:Documentation]

Release Note:
Improved Contributing Guidelines